# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Agzos Hub internal management system API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /sites:
    get:
      operationId: listSites
      tags: [sites]
      summary: List all sites
      parameters:
        - name: status
          in: query
          schema:
            type: string
            enum: [active, development, maintenance, paused]
        - name: search
          in: query
          schema:
            type: string
      responses:
        "200":
          description: List of sites
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Site"
    post:
      operationId: createSite
      tags: [sites]
      summary: Create a new site
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateSiteBody"
      responses:
        "201":
          description: Created site
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Site"
  /sites/{id}:
    get:
      operationId: getSite
      tags: [sites]
      summary: Get site by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Site
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Site"
    put:
      operationId: updateSite
      tags: [sites]
      summary: Update site
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateSiteBody"
      responses:
        "200":
          description: Updated site
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Site"
    delete:
      operationId: deleteSite
      tags: [sites]
      summary: Delete site
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
...
  /projects:
    get:
      operationId: listProjects
      tags: [projects]
      summary: List all projects
      parameters:
        - name: status
          in: query
          schema:
            type: string
            enum: [planning, active, review, completed, paused, cancelled]
        - name: clientId
          in: query
          schema:
            type: integer
        - name: search
          in: query
          schema:
            type: string
      responses:
        "200":
          description: List of projects
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Project"
    post:
      operationId: createProject
      tags: [projects]
      summary: Create a new project
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateProjectBody"
      responses:
        "201":
          description: Created project
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Project"
  /projects/{id}:
    get:
      operationId: getProject
      tags: [projects]
      summary: Get project by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Project
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Project"
    put:
      operationId: updateProject
      tags: [projects]
      summary: Update project
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateProjectBody"
      responses:
        "200":
          description: Updated project
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Project"
    delete:
      operationId: deleteProject
      tags: [projects]
      summary: Delete project
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
...
  /tasks:
    get:
      operationId: listTasks
      tags: [tasks]
      summary: List tasks
      parameters:
        - name: projectId
          in: query
          schema:
            type: integer
        - name: assigneeId
          in: query
          schema:
            type: integer
        - name: status
          in: query
          schema:
            type: string
            enum: [todo, in_progress, review, done]
      responses:
        "200":
          description: List of tasks
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Task"
    post:
      operationId: createTask
      tags: [tasks]
      summary: Create a task
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateTaskBody"
      responses:
        "201":
          description: Created task
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Task"
  /tasks/{id}:
    put:
      operationId: updateTask
      tags: [tasks]
      summary: Update task
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateTaskBody"
      responses:
        "200":
          description: Updated task
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Task"
    delete:
      operationId: deleteTask
      tags: [tasks]
      summary: Delete task
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  # Clients
...
  /team:
    get:
      operationId: listTeamMembers
      tags: [team]
      summary: List team members
      responses:
        "200":
          description: List of team members
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/TeamMember"
    post:
      operationId: createTeamMember
      tags: [team]
      summary: Invite a team member
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateTeamMemberBody"
      responses:
        "201":
          description: Created team member
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TeamMember"
  /team/{id}:
    put:
      operationId: updateTeamMember
      tags: [team]
      summary: Update team member
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateTeamMemberBody"
      responses:
        "200":
          description: Updated team member
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TeamMember"
    delete:
      operationId: deleteTeamMember
      tags: [team]
      summary: Remove team member
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  # Financial
...
components:
  schemas:
...
    Site:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        url:
          type: string
        status:
          type: string
          enum: [active, development, maintenance, paused]
        clientId:
          type: integer
        clientName:
          type: string
        adminUrl:
          type: string
        platform:
          type: string
        lastDeployment:
          type: string
          format: date-time
        monthlyVisitors:
          type: integer
        notes:
          type: string
        createdAt:
          type: string
          format: date-time
      required: [id, name, url, status, createdAt]
    CreateSiteBody:
      type: object
      properties:
        name:
          type: string
        url:
          type: string
        status:
          type: string
          enum: [active, development, maintenance, paused]
        clientId:
          type: integer
        adminUrl:
          type: string
        platform:
          type: string
        notes:
          type: string
      required: [name, url, status]
    UpdateSiteBody:
      type: object
      properties:
        name:
          type: string
        url:
          type: string
        status:
          type: string
          enum: [active, development, maintenance, paused]
        clientId:
          type: integer
        adminUrl:
          type: string
        platform:
          type: string
        notes:
          type: string
...
    Project:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        description:
          type: string
        status:
          type: string
          enum: [planning, active, review, completed, paused, cancelled]
        clientId:
          type: integer
        clientName:
          type: string
        assigneeId:
          type: integer
        assigneeName:
          type: string
        budget:
          type: number
        startDate:
          type: string
          format: date
        dueDate:
          type: string
          format: date
        completedAt:
          type: string
          format: date-time
        progress:
          type: integer
          minimum: 0
          maximum: 100
        priority:
          type: string
          enum: [low, medium, high, urgent]
        createdAt:
          type: string
          format: date-time
      required: [id, name, status, priority, progress, createdAt]
    CreateProjectBody:
      type: object
      properties:
        name:
          type: string
        description:
          type: string
        status:
          type: string
          enum: [planning, active, review, completed, paused, cancelled]
        clientId:
          type: integer
        assigneeId:
          type: integer
        budget:
          type: number
        startDate:
          type: string
          format: date
        dueDate:
          type: string
          format: date
        priority:
          type: string
          enum: [low, medium, high, urgent]
      required: [name, status, priority]
    UpdateProjectBody:
      type: object
      properties:
        name:
          type: string
        description:
          type: string
        status:
          type: string
          enum: [planning, active, review, completed, paused, cancelled]
        clientId:
          type: integer
        assigneeId:
          type: integer
        budget:
          type: number
        startDate:
          type: string
          format: date
        dueDate:
          type: string
          format: date
        priority:
          type: string
          enum: [low, medium, high, urgent]
        progress:
          type: integer
...
    Task:
      type: object
      properties:
        id:
          type: integer
        title:
          type: string
        description:
          type: string
        status:
          type: string
          enum: [todo, in_progress, review, done]
        projectId:
          type: integer
        projectName:
          type: string
        assigneeId:
          type: integer
        assigneeName:
          type: string
        priority:
          type: string
          enum: [low, medium, high, urgent]
        dueDate:
          type: string
          format: date
        createdAt:
          type: string
          format: date-time
      required: [id, title, status, priority, createdAt]
    CreateTaskBody:
      type: object
      properties:
        title:
          type: string
        description:
          type: string
        status:
          type: string
          enum: [todo, in_progress, review, done]
        projectId:
          type: integer
        assigneeId:
          type: integer
        priority:
          type: string
          enum: [low, medium, high, urgent]
        dueDate:
          type: string
          format: date
      required: [title, status, priority]
    UpdateTaskBody:
      type: object
      properties:
        title:
          type: string
        description:
          type: string
        status:
          type: string
          enum: [todo, in_progress, review, done]
        projectId:
          type: integer
        assigneeId:
          type: integer
        priority:
          type: string
          enum: [low, medium, high, urgent]
        dueDate:
          type: string
          format: date
...
    TeamMember:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        email:
          type: string
        role:
          type: string
          enum: [admin, account_manager, traffic_manager, designer, developer, financial, client]
        status:
          type: string
          enum: [active, invited, inactive]
        avatarUrl:
          type: string
        activeProjects:
          type: integer
        joinedAt:
          type: string
          format: date-time
      required: [id, name, email, role, status]
    CreateTeamMemberBody:
      type: object
      properties:
        name:
          type: string
        email:
          type: string
        role:
          type: string
          enum: [admin, account_manager, traffic_manager, designer, developer, financial, client]
      required: [name, email, role]
    UpdateTeamMemberBody:
      type: object
      properties:
        name:
          type: string
        role:
          type: string
          enum: [admin, account_manager, traffic_manager, designer, developer, financial, client]
        status:
          type: string
          enum: [active, invited, inactive]
...